dashboard

+++ app/Http/Controllers/Dashboard/DoctorController.php
@@ -9,6 +9,7 @@ use Illuminate\Http\Request;
 use App\Models\DoctorInfo;
 use App\Models\Specialty;
 use App\Models\User;
+use Illuminate\Support\Facades\Redirect;
 
 class DoctorController extends Controller
 {
@@ -23,8 +24,9 @@ class DoctorController extends Controller
      */
     public function show($id)
     {
-        $user = User::findOrFail($id);
-        return view('dashboard.doctor.edit', compact('user'));
+        $user = User::find($id);
+        $allSpecialties = Specialty::get();
+        return view('dashboard.doctor.edit', compact('user', 'allSpecialties'));
     }
 
     /**
@@ -33,7 +35,7 @@ class DoctorController extends Controller
     public function create()
     {
         $allSpecialties = Specialty::get();
-        return view('dashboard.doctor.add', compact('allSpecialties'));
+        return view('dashboard.doctor.create', compact('allSpecialties'));
     }
 
     /**
@@ -41,35 +43,30 @@ class DoctorController extends Controller
      */
     public function store(Request $request)
     {
-        $validation = Validator::make($request->all(), [
+        $request->validate([
             'name' => 'required|string',
             'email' => 'required|email|unique:users',
             'phone' => 'required|unique:users',
-            'specialty_id' => 'required|exits:specialties,id',
-            'password' => 'required|string|confirmed'
+            'specialty_id' => 'required|exists:specialties,id',
+            'password' => 'required',
+            'confirm_password' => 'required|same:password'
         ]);
 
-        if ($validation->fails()) {
-            session()->flash('error', 'Sorry! Try Again.' . $validation->errors()->first());
-            return redirect()->back();
-        }
         $user = User::create([
             'name' => $request->name,
             'email' => $request->email,
             'phone' => $request->phone,
-            'user_type' => '1',
+            'user_type' => '0',
             'password' => Hash::make($request->password),
         ]);
-        if ($user->user_type == 0) {
-            DoctorInfo::create([
-                'user_id' => $user->id,
-                'specialty_id' => $request->specialty_id,
-                'bio' => 'Don’t fear any illness…everything has its cure.',
-                'fees' => 0,
-            ]);
-        }
+        DoctorInfo::create([
+            'user_id' => $user->id,
+            'specialty_id' => $request->specialty_id,
+            'bio' => 'Don’t fear any illness…everything has its cure.',
+            'fees' => 0,
+        ]);
         notify()->success('You are awesome, The Doctor account has been created successfull!');
-        return redirect()->route('patients.index');
+        return redirect()->route('doctors.index');
     }
 
     /**
@@ -78,16 +75,8 @@ class DoctorController extends Controller
     public function search(Request $request)
     {
         $search = $request->get('search');
-        $allDoctors = User::where('name', 'like', '%' . $search . '%')->where('user_type', 2)->paginate(10);
-        return view('dashboard.user.doctors', compact('allDoctors'));
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
+        $allDoctors = User::where('name', 'like', '%' . $search . '%')->where('user_type', "0")->paginate(10);
+        return view('dashboard.doctor.manage', compact('allDoctors'));
     }
 
     /**
@@ -95,18 +84,49 @@ class DoctorController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        //
-    }
+        $user = User::where('id', $id)->first();
+        if ($user) {
+            $request->validate([
+                'name' => 'sometimes|required|string',
+                'email' => 'sometimes|required|email|unique:users,email,' . $user->id,
+                'phone' => 'sometimes|required|unique:users,phone,' . $user->id,
+                'specialty_id' => 'sometimes|required|exists:specialties,id'
+            ]);
 
+            $user->name = $request->name ?? $user->name;
+            $user->email = $request->email ?? $user->email;
+            $user->phone = $request->phone ?? $user->phone;
+            if (!empty($request->password)) {
+                $request->validate([
+                    'password' => 'sometimes|required',
+                    'confirm_password' => 'sometimes|required|same:password'
+                ]);
+                $user->password = Hash::make($request->password);
+            }
+            if (!empty($request->specialty_id)) {
+                $user->doctor_info->specialty_id = $request->specialty_id ?? $user->doctor_info->specialty_id;
+            }
+            $user->save();
+            notify()->success('You are awesome, The Doctor account has been created successfull!');
+            return redirect()->route('doctors.index');
+        } else {
+            notify()->error('Opps!');
+            return redirect()->back();
+        }
+    }
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(string $id)
     {
-        $doctor = User::findOrFail($id);
-        $doctor->delete();
-
-        return redirect()->route('Doctors')->with('success', 'Done');
+        $doctor = User::find($id);
+        if ($doctor) {
+            $doctor->delete();
+            notify()->success('You are awesome, The Doctor account has been deleted successfull!');
+        } else {
+            notify()->error('Opps!, The Doctor account has been deleted before');
+        }
+        return redirect()->back();
     }
 }
